+++ packages/core/src/index.ts
@@ -56,4 +56,4 @@ export {
   validateSession,
 } from "./utils/session";
 
-export { fetchPools } from "./magic-swap";
+export { fetchPools } from "./magicswap";
