+++ packages/core/src/magicswap/graph-queries.ts
