+++ packages/core/src/magicswap/types.ts
